Fix has_module for non-builtin modules
When py.test is invoked with the --ansible-module-path flag, this flag
is correctly passed down to the underlying module, but the Ansible
module loader is never informed that this module path should be
searched. Tell Anisble about it so it gives us correct results and
doesn't prevent us from executing modules that otherwise would succeed.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>